requirements.txt
django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
PyJWT==2.8.0
python-dotenv==1.0.0
django-filter==23.5
djangorestframework-simplejwt==5.3.0
setuptools==69.0.3
markdown==3.5.1
coreapi==2.3.3
uritemplate==4.1.1
